# backend/docs/swagger/swagger.yaml
basePath: /
...
host: localhost:8088
info:
  contact: {}
  description: Bilingual CMS backend API for Impress (印迹). Supports content management,
    articles, pages, themes, media, and more.
  title: Impress CMS API
  version: "1.0"
paths:
...
  /admin/menus/{id}/items:
    post:
      consumes:
      - application/json
      description: Create a new menu item within a menu group
      parameters:
      - description: Menu group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Menu item data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create menu item
      tags:
      - Menus
  /admin/menus/{id}/items/{itemId}:
    delete:
      description: Delete a menu item by ID
      parameters:
      - description: Menu group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Menu item ID
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete menu item
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: Update a menu item's properties
      parameters:
      - description: Menu group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Menu item ID
        in: path
        name: itemId
        required: true
        type: integer
      - description: Updated menu item data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update menu item
      tags:
      - Menus
...
  /admin/search/rebuild:
    post:
      description: Rebuild the full-text search index from scratch
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Rebuild search index
      tags:
      - Search (Admin)
  /admin/seo/robots:
    get:
      description: Returns the robots.txt content for admin editing
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              content:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get robots.txt (admin)
      tags:
      - SEO
    put:
      consumes:
      - application/json
      description: Update the robots.txt content
      parameters:
      - description: Robots.txt content
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              content:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update robots.txt
      tags:
      - SEO
...
  /public/search:
    get:
      description: Full-text search across articles and pages
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: zh
        description: Locale (zh or en)
        in: query
        name: locale
        type: string
      - description: Content type filter
        in: query
        name: type
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Search content
      tags:
      - Search
  /public/search/suggest:
    get:
      description: Returns autocomplete suggestions for a search prefix
      parameters:
      - description: Search prefix
        in: query
        name: q
        required: true
        type: string
      - default: zh
        description: Locale (zh or en)
        in: query
        name: locale
        type: string
      - default: 5
        description: Max suggestions
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: Search suggestions
      tags:
      - Search
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" for JWT authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"